CartActivity bug fixed

+++ StuffHub/app/src/main/java/com/example/stuffhub/CartActivity.java
@@ -37,7 +37,7 @@ public class CartActivity extends AppCompatActivity {
     private Button NextProcessBtn;
     private TextView txtTotalAmount, txtMsg1;
 
-    private int overTotalPrice = 0;
+    private float overTotalPrice = 0;
 
 
     @Override
@@ -93,7 +93,7 @@ public class CartActivity extends AppCompatActivity {
                holder.TxtProductPrice.setText("Price = $"+model.getPrice()+ " Pesos");
                holder.TxtProductName.setText("Name = $"+model.getPname());
 
-               int oneTypeProductPrice = ((Integer.valueOf(model.getPrice()))) * Integer.valueOf(model.getQuantity());
+               float oneTypeProductPrice = ((Float.valueOf(model.getPrice()))) * Float.valueOf(model.getQuantity());
                overTotalPrice = overTotalPrice + oneTypeProductPrice;
 
                holder.itemView.setOnClickListener(new View.OnClickListener() {
